+++ browser/i18n/gl_GL.js
@@ -102,6 +102,6 @@ module.exports = {
         "Area": "Nunaminertaq",
         "Distance/Radius": "Ungasissuseq/tukimut takissuseq",
         "Type": "Iluseq",
-
+        "select-zip-shape":"Nakorsinnaavoq .zip fil attaveq",
     }
 };